requirements.txt
torch
torchvision
numpy
matplotlib
scikit-image
scikit-learn
pillow
seaborn
tensorboard
